perf(tts/styletts2): iteration_3 — mixed-precision build (7 fp16 / 1 fp32)
Flip _STAGE_PRECISION defaults to fp16 for every stage that A/B'd clean
against the iteration_2 fp32 baseline; keep fused_f0n_har_source at fp32
because the har_source half computes sin(2π · cumsum(f0)) over 88 200
samples and fp16 cumsum drifts ~10 bits, producing audible phase
distortion in the second half of the clip.

Final precision matrix:

  text_encoder, bert, ref_encoder, fused_diffusion_sampler,
  duration_predictor, decoder_pre, decoder_upsample          → fp16
  fused_f0n_har_source                                       → fp32

Verified end-to-end via four sanity WAVs:

  sanity_fp16_mixed.wav        (5 fp16)  clean
  sanity_fp16_plus_decpre.wav  (6 fp16)  clean
  sanity_fp16_plus_decup.wav   (7 fp16)  clean    ← this build
  sanity_fp16_plus_f0n.wav     (8 fp16)  degraded second half

Standalone fp16 vs fp32 warm bench (M-series Mac):

  fused_diffusion_sampler / ALL        18.3 → 14.7 ms  (-3.6 ms)
  fused_f0n_har_source / CPU_ONLY      11.9 →  7.2 ms  (kept fp32 anyway)
  decoder_pre                          35   →  7   ms  (-28 ms)
  decoder_upsample / CPU_ONLY          ~600 → 285  ms  (-315 ms)

Pipeline-stage sum, 3-iter cool sweep:

  iteration_2 fp32        min 782   avg 898   max 1075 ms
  iteration_3             min 460   avg 683   max 1110 ms
                                  -322ms        -215ms

Disk: 514 → 274 MB (-240 MB / -47 %).

decoder_upsample at fp16 had been flagged as a +240 ms regression in
earlier trials when placed on ALL; on the Trial 8b CPU_ONLY placement
the fp16 build is dominantly faster, so the regression doesn't apply.

Adds iteration_2/README.md (the previous fp32 baseline) and
iteration_3/README.md describing the mixed-precision build, drift
rationale, and per-stage timings.

+++ models/tts/styletts2/coreml/inference.py
@@ -116,21 +116,22 @@ _STAGE_COMPUTE: dict[str, ct.ComputeUnit] = {
 # for the legacy fp16 path (only valid for stages that have an
 # `<stage>_fp16.mlpackage`).
 _STAGE_PRECISION: dict[str, str] = {
-    "text_encoder":             "fp32",
-    "bert":                     "fp32",
-    "ref_encoder":              "fp32",
-    "fused_diffusion_sampler":  "fp32",
-    "diffusion_unet":           "fp32",
-    "duration_predictor":       "fp32",
+    "text_encoder":             "fp16",
+    "bert":                     "fp16",
+    "ref_encoder":              "fp16",
+    "fused_diffusion_sampler":  "fp16",
+    "diffusion_unet":           "fp32",  # legacy fallback (use --no-fused)
+    "duration_predictor":       "fp16",
+    # fused_f0n_har_source must stay fp32: the har_source half computes
+    # sin(2π × cumsum(f0)) at audio rate (88 200 samples). fp16 cumsum
+    # drifts ~10 bits over that span and produces audible phase
+    # distortion in the second half of the clip. Verified by A/B in
+    # iteration_3 sweep.
     "fused_f0n_har_source":     "fp32",
-    "f0n_predictor":            "fp32",
-    # har_source must stay fp32: it computes sin(2π × cumsum(f0)) at
-    # audio rate (74 400 samples). fp16 cumsum drifts ~10 bits over that
-    # span and produces audible phase distortion in the second half of
-    # the clip.
-    "har_source":               "fp32",
-    "decoder_pre":              "fp32",
-    "decoder_upsample":         "fp32",
+    "f0n_predictor":            "fp32",  # legacy fallback
+    "har_source":               "fp32",  # legacy fallback (same drift)
+    "decoder_pre":              "fp16",
+    "decoder_upsample":         "fp16",
 }
